requirements.txt
numpy==2.0.1
pandas==2.2.2
xlsx2csv==0.8.3
selenium==4.23.1
gspread==5.12.4
gspread-dataframe==4.0.0
python-dotenv==1.0.1
jinja2==3.1.4
pyyaml==6.0.2
./2025_03_WebHandler
